+++ app/src/main/java/com/example/mealtracker/foodDetails/NutrientsX.kt
@@ -2,7 +2,7 @@ package com.example.mealtracker.foodDetails
 
 data class NutrientsX(
     val CHOCDF: Double,
-    val ENERC_KCAL: Int,
+    val ENERC_KCAL: Double,
     val FAT: Double,
     val FIBTG: Double,
     val PROCNT: Double
